Add files via upload
Added homework 4. Problem one asks us to create a boolean satisfiability solver, given psuedocode for the DPLL algorithm. Problem two asks us to use the SAT solver from problem one to determine the 3-colorability (and colors if possible) for an undirected graph. This latter question is a illustration of the power of translation between NPC problems.